Reorg form manager edit components relative to where they are used
