+++ packages/design/src/FormManager/FormEdit/DraggableList.tsx
@@ -18,49 +18,6 @@ import {
 import { CSS } from '@dnd-kit/utilities';
 import { useFormEditStore } from './store';
 
-const SortableItem = ({
-  id,
-  children,
-}: {
-  id: UniqueIdentifier;
-  children: React.ReactNode;
-}) => {
-  const { attributes, listeners, setNodeRef, transform, transition } =
-    useSortable({ id });
-  const context = useFormEditStore(state => state.context);
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={{
-        transform: CSS.Transform.toString(transform),
-        transition,
-      }}
-    >
-      <div className="grid-row grid-gap">
-        <div
-          className="grid-col-12 grid-col width-full"
-          {...listeners}
-          {...attributes}
-          style={{ cursor: 'grab' }}
-        >
-          <svg
-            className="usa-icon margin-x-auto display-block"
-            aria-hidden="true"
-            focusable="false"
-            role="img"
-          >
-            <use
-              xlinkHref={`${context.uswdsRoot}img/sprite.svg#drag_handle`}
-            ></use>
-          </svg>
-        </div>
-        <div className="grid-col-12 grid-col">{children}</div>
-      </div>
-    </div>
-  );
-};
-
 type DraggableListProps = React.PropsWithChildren<{
   order: UniqueIdentifier[];
   updateOrder: (order: UniqueIdentifier[]) => void;
@@ -108,4 +65,45 @@ export const DraggableList: React.FC<DraggableListProps> = ({
   );
 };
 
-export default DraggableList;
+const SortableItem = ({
+  id,
+  children,
+}: {
+  id: UniqueIdentifier;
+  children: React.ReactNode;
+}) => {
+  const { attributes, listeners, setNodeRef, transform, transition } =
+    useSortable({ id });
+  const context = useFormEditStore(state => state.context);
+
+  return (
+    <div
+      ref={setNodeRef}
+      style={{
+        transform: CSS.Transform.toString(transform),
+        transition,
+      }}
+    >
+      <div className="grid-row grid-gap">
+        <div
+          className="grid-col-12 grid-col width-full"
+          {...listeners}
+          {...attributes}
+          style={{ cursor: 'grab' }}
+        >
+          <svg
+            className="usa-icon margin-x-auto display-block"
+            aria-hidden="true"
+            focusable="false"
+            role="img"
+          >
+            <use
+              xlinkHref={`${context.uswdsRoot}img/sprite.svg#drag_handle`}
+            ></use>
+          </svg>
+        </div>
+        <div className="grid-col-12 grid-col">{children}</div>
+      </div>
+    </div>
+  );
+};
