+++ packages/design/src/FormManager/FormEdit/PreviewSequencePattern.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 
 import { PatternProps, getPattern } from '@atj/forms';
+import { SequencePattern } from '@atj/forms/src/patterns/sequence';
 
 import { PatternComponent } from '../../Form';
-import DraggableList from './DraggableList';
+
+import { DraggableList } from './DraggableList';
 import { useFormEditStore } from './store';
-import { SequencePattern } from '@atj/forms/src/patterns/sequence';
 
 // TODO: consider merging this component with DraggableList, to clean up
 // sematics around how its children are handled.
