+++ packages/design/src/FormManager/FormEdit/components/FormSummaryEdit.tsx
@@ -3,9 +3,13 @@ import React from 'react';
 import { type FormSummaryProps, type PatternId } from '@atj/forms';
 
 import FormSummary from '../../../Form/components/FormSummary';
-import { PatternEditForm, usePatternEditFormContext } from '../PatternEditForm';
 import { useIsPatternSelected } from '../store';
 import { PatternEditComponent } from '../types';
+
+import {
+  PatternEditForm,
+  usePatternEditFormContext,
+} from './common/PatternEditForm';
 
 const FormSummaryEdit: PatternEditComponent<FormSummaryProps> = props => {
   const isSelected = useIsPatternSelected(props.previewProps._patternId);
