+++ packages/design/src/FormManager/FormEdit/index.tsx
@@ -1,10 +1,13 @@
 import React, { useEffect } from 'react';
 
-import { Blueprint } from '@atj/forms';
+import { Blueprint, createFormSession } from '@atj/forms';
 import { type FormService } from '@atj/form-service';
 
-import { AddPattern } from './AddPattern';
-import { PreviewForm } from './Preview';
+import Form, { ComponentForPattern, PatternComponent } from '../../Form';
+
+import { AddPatternDropdown } from './AddPatternDropdown';
+import { PreviewPattern } from './PreviewPattern';
+import { PatternPreviewSequence } from './PreviewSequencePattern';
 import { FormEditProvider, useFormEditStore } from './store';
 import { type FormEditUIContext } from './types';
 
@@ -36,6 +39,8 @@ export default function FormEdit({
 
 const EditForm = ({ saveForm }: { saveForm: (form: Blueprint) => void }) => {
   const { form } = useFormEditStore();
+  const uiContext = useFormEditStore(state => state.context);
+  const disposable = createFormSession(form); // nullSession instead?
   useEffect(() => {
     saveForm(form);
   }, [form]);
@@ -44,14 +49,41 @@ const EditForm = ({ saveForm }: { saveForm: (form: Blueprint) => void }) => {
     <div className="position-relative">
       <div className="grid-row">
         <div className="grid-col-12">
-          <AddPattern />
+          <AddPatternDropdown />
         </div>
       </div>
       <div className="grid-row">
         <div className="grid-col-12">
-          <PreviewForm />
+          <Form
+            isPreview={true}
+            context={{
+              config: uiContext.config,
+              // TODO: We might want to hoist this definition up to a higher level,
+              // so we don't have to regenerate it every time we render the form.
+              components: createPreviewComponents(uiContext.components),
+              uswdsRoot: uiContext.uswdsRoot,
+            }}
+            session={disposable}
+          ></Form>
         </div>
       </div>
     </div>
   );
+};
+
+const createPreviewComponents = (
+  components: ComponentForPattern
+): ComponentForPattern => {
+  const previewComponents: ComponentForPattern = {};
+  // TODO: Create a configurable way to to define preview components.
+  for (const [patternType, Component] of Object.entries(components)) {
+    previewComponents[patternType] = Component;
+    if (patternType === 'sequence') {
+      previewComponents[patternType] =
+        PatternPreviewSequence as PatternComponent;
+    } else {
+      previewComponents[patternType] = PreviewPattern;
+    }
+  }
+  return previewComponents;
 };
