+++ packages/design/src/FormManager/FormManagerLayout/BottomNavigation.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const BottomNavigation = () => {
+  return (
+    <div className="position-sticky bottom-0 z-top bg-white padding-1">
+      Bottom navigation
+    </div>
+  );
+};
